Extract out `RBSCommentsToSorbetSigs` module
